// ts-server/src/types/models.types.ts
import mongoose from 'mongoose'

export type UserDocument = mongoose.Document & {
  _id: string
  name: string
  email: string
  password: string
  username: string
  date_of_birth: string
  created_at: string
  updated_at: string
  friends_count: number
  followers_count: number
  bio: string
  location: string
  img: string
  header_image: string
  links: string[]
}